Separated tests in PruningTest.java to PruningNoFilesTest.java, PruningOneNewFileTest.java, PruningOneOldFileTest.java, PruningOneOldOneNewFileTest.java, PruningTwoNewFilesTest.java and PruningTwoOldFilesTest.java
